moved common tests from test_mini_formation into test_common.py
